Upgrade font awesome. Add money gem. Update migrations

+++ db/migrate/20140602230227_create_routes.rb
@@ -3,10 +3,10 @@ class CreateRoutes < ActiveRecord::Migration
     create_table :routes do |t|
       t.date :date
       t.integer :trip_number
-      t.float :advance
-      t.float :reimbursement
-      t.float :gross_revenue
-      t.float :net_revenue
+      t.money :advance
+      t.money :reimbursement
+      t.money :gross_revenue
+      t.money :net_revenue
       t.references :driver, index:true
 
       t.timestamps
