+++ db/migrate/20140609014319_add_trailer_rental_to_routes.rb
@@ -1,6 +1,6 @@
 class AddTrailerRentalToRoutes < ActiveRecord::Migration
   def change
     add_column :routes, :rented_trailer, :boolean
-    add_column :routes, :trailer_rental, :float
+    add_column :routes, :trailer_rental, :money
   end
 end
